Working Cli again
