+++ src/cli.rs
@@ -1,3 +1,4 @@
+use std::fs;
 use std::io::{self, stdin, BufRead};
 use std::path::PathBuf;
 
@@ -5,6 +6,9 @@ use bf_core::{BFInterpreter, BFProgram};
 use clap::builder::styling::AnsiColor;
 use clap::builder::Styles;
 use clap::{Parser, Subcommand};
+
+use crate::compiler::BrainCrabCompiler;
+use crate::parser::BrainCrabParser;
 
 fn get_cli_style() -> Styles {
     Styles::styled()
@@ -16,13 +20,30 @@ fn get_cli_style() -> Styles {
 
 #[derive(Parser)]
 #[command(version, about, long_about, styles=get_cli_style())]
-pub(crate) struct Cli {
+pub struct Cli {
     #[command(subcommand)]
     command: Commands,
 }
 
 #[derive(Subcommand)]
 enum Commands {
+    /// Compile a BrainCrab script to Brainfuck.
+    Compile {
+        path: PathBuf,
+        #[arg(short, long)]
+        output: Option<PathBuf>,
+    },
+
+    /// Run a BrainCrab script as Brainfuck.
+    Run { path: PathBuf },
+
+    /// BF Commands
+    #[command(subcommand)]
+    BF(BFCommands),
+}
+
+#[derive(Subcommand)]
+enum BFCommands {
     /// Run a Brainfuck file.
     Run { path: PathBuf },
 
@@ -34,14 +55,77 @@ enum Commands {
 }
 
 impl Cli {
-    pub(crate) fn start(self) -> io::Result<()> {
+    pub fn start(self) -> io::Result<()> {
         match self.command {
             Commands::Run { path } => Self::run(path),
-            Commands::Repl => Self::repl(),
+            Commands::Compile { path, output } => Self::compile(path, output),
+            Commands::BF(BFCommands::Run { path }) => Self::bf_run(path),
+            Commands::BF(BFCommands::Repl) => Self::bf_repl(),
         }
     }
 
     fn run(path: PathBuf) -> io::Result<()> {
+        let script = fs::read_to_string(&path)?;
+        let mut parser = BrainCrabParser::new();
+        let parse_result = parser.parse_program(&script);
+
+        match parse_result {
+            Ok(parsed) => {
+                let program = parsed.value;
+                let compiled_abf = BrainCrabCompiler::compile_abf(program);
+                match compiled_abf {
+                    Ok(compiled_abf) => {
+                        let bf = compiled_abf.to_bf();
+                        let mut interpreter = BFInterpreter::new();
+                        interpreter.run(&bf);
+                    }
+                    Err(error) => {
+                        eprintln!("Encountered error while compiling {path:?}:");
+                        eprintln!("{error:?}");
+                    }
+                }
+            }
+            Err(error) => {
+                eprintln!("Encountered error while parsing {path:?}:");
+                eprintln!("{error}")
+            }
+        }
+        Ok(())
+    }
+
+    fn compile(path: PathBuf, output: Option<PathBuf>) -> io::Result<()> {
+        let script = fs::read_to_string(&path)?;
+        let mut parser = BrainCrabParser::new();
+        let parse_result = parser.parse_program(&script);
+
+        match parse_result {
+            Ok(parsed) => {
+                let program = parsed.value;
+                let compiled_abf = BrainCrabCompiler::compile_abf(program);
+                match compiled_abf {
+                    Ok(compiled_abf) => {
+                        let bf = compiled_abf.to_bf().to_string();
+                        if let Some(output_path) = output {
+                            fs::write(output_path, bf)?;
+                        } else {
+                            println!("{bf}");
+                        }
+                    }
+                    Err(error) => {
+                        eprintln!("Encountered error while compiling {path:?}:");
+                        eprintln!("{error:?}");
+                    }
+                }
+            }
+            Err(error) => {
+                eprintln!("Encountered error while parsing {path:?}:");
+                eprintln!("{error}")
+            }
+        }
+        Ok(())
+    }
+
+    fn bf_run(path: PathBuf) -> io::Result<()> {
         let script = std::fs::read_to_string(path)?;
         let program = BFProgram::parse(&script).expect("Invalid program");
         let mut interpreter = BFInterpreter::new();
@@ -49,7 +133,7 @@ impl Cli {
         Ok(())
     }
 
-    fn repl() -> io::Result<()> {
+    fn bf_repl() -> io::Result<()> {
         let mut interpreter = BFInterpreter::new();
         loop {
             let mut buffer = String::new();
