+++ src/lib.rs
@@ -1,7 +1,7 @@
 mod absolute_bf;
 mod allocator;
-mod ast;
-mod cli;
+pub mod ast;
+pub mod cli;
 pub mod compiler;
 pub mod parser;
 mod value;
